--- src/models/types.ts
@@ -1,5 +1,0 @@
-export type Message = {
-    text: string;
-    clientId: string;
-    created_time: string;
-};
